# Description



<!-- Add a description of the changes that this PR introduces and the files that
are the most critical to review. -->


These changes use b.Loop() to simplify the code and improve performance

Supported by Go Team, more info: https://go.dev/blog/testing-b-loop

More info can see https://go.dev/issue/73137.

Before this change:

```shell
go test -run=^$ -bench=. ./x/staking -timeout=1h 
goos: darwin
goarch: arm64
pkg: github.com/cosmos/cosmos-sdk/x/staking
cpu: Apple M4
BenchmarkValidateGenesis10Validators-10          	 2375758	       501.9 ns/op	    1008 B/op	      18 allocs/op
BenchmarkValidateGenesis100Validators-10         	  356570	      3351 ns/op	    6929 B/op	     108 allocs/op
BenchmarkValidateGenesis400Validators-10         	   88380	     13993 ns/op	   26644 B/op	     408 allocs/op
BenchmarkApplyAndReturnValidatorSetUpdates-10    	    3522	    324170 ns/op
--- BENCH: BenchmarkApplyAndReturnValidatorSetUpdates-10
    keeper_bench_test.go:45: vals count: 150
PASS
ok  	github.com/cosmos/cosmos-sdk/x/staking	6.171s
```

After:


```shell
go test -run=^$ -bench=. ./x/staking -timeout=1h 
goos: darwin
goarch: arm64
pkg: github.com/cosmos/cosmos-sdk/x/staking
cpu: Apple M4
BenchmarkValidateGenesis10Validators-10          	 2533317	       470.4 ns/op	     816 B/op	      17 allocs/op
BenchmarkValidateGenesis100Validators-10         	  348805	      3314 ns/op	    6737 B/op	     107 allocs/op
BenchmarkValidateGenesis400Validators-10         	   89534	     13380 ns/op	   26452 B/op	     407 allocs/op
BenchmarkApplyAndReturnValidatorSetUpdates-10    	    3540	    324747 ns/op
--- BENCH: BenchmarkApplyAndReturnValidatorSetUpdates-10
    keeper_bench_test.go:45: vals count: 150
    keeper_bench_test.go:45: vals count: 150
    keeper_bench_test.go:45: vals count: 150
PASS
ok  	github.com/cosmos/cosmos-sdk/x/staking	5.888s
```

<!-- This is an auto-generated comment: release notes by coderabbit.ai -->

## Summary by CodeRabbit

* **Tests**
  * Updated staking module benchmark tests to use the latest benchmarking loop mechanism for more consistent measurements. No changes to product behavior.
* **Chores**
  * General test maintenance to align with current testing practices.
* **Note**
  * No user-facing changes or feature updates in this release.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->